pyproject.toml
[project]
name = "causalitynet"
version = "0.1.0"
description = "CausalityNet using crewAI"
authors = [{ name = "Abdul Haque",email = "abdulhaque.dev@gmail.com"}, 
           { name = "Umm e Hani" ,email = "umme.dev@gmail.com"}]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "0.108.0"
unstructured = "==0.17.2"
tools = "^0.1.9"
python-dotenv = "1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
google = ">=3.0.0"
numpy = "^1.24.0"
tqdm = "^4.64.0"
ipython = "^8.12.0"
pandas = "^1.5.3"
scikit-learn = "^1.3.0"
langchain_community = "0.3.20"
neo4j = "5.28.1"
fastembed = "0.6.0"
langchain_neo4j = "0.4.0"
onnxruntime = "1.17.0"
tavily-python = "0.5.3"

[project.scripts]
kickoff = "causalitynet.main:kickoff"
plot = "causalitynet.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"